requirements.txt
Flask==2.3.3
gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
secure-smtplib==0.1.1
Werkzeug==2.3.7
redis==5.0.1
rq==1.15.1
email-validator==2.1.0.post1